Add files via upload
The application relies on NodeJS and Bower For UI. Zurb Foundation 6.3.1 is the CSS template used for the project alongside some very help fully JavaScript dependencies such as PACE and TEXTILLATE. In Order to work with the project it will be require integrating the Foundation 6.3.1 template or in case you need to use the template of your choice, there you would have to modify the Webpack script to suit your demands. The project is quite very for it was belt mainly for further development with emphasis on the UI. If you would need any guide, a README file will be attached to the project.